% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cow_Calf_Sub_Model.R
\name{cow.calf}
\alias{cow.calf}
\title{model bacteria generation from cow-calf agricultural operation}
\usage{
cow.calf(chr.file.input)
}
\arguments{
\item{chr.file.input}{is the input file for the model}
}
\description{
This function is the bacteria source-model cow-calf systems
and generates input for HSPF. The specific outputs from
this source model are loads from the cow-calf system to the land
and directly to the stream. The load to the land is in the form
of load/acre for each PLS in a sub-watershed that the source-model
contributes to and the daily load to the stream.
The input for the model is from an ASCII text file. For the input
file. The symbol used for comments in the input file is "***".
":" is used to seperate parameter names or descriptions from values.
The function returns a data.frame with populations for different
locations in the sub-watershed along with the associated bacteria loads.
}
