% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-get-HSPF-sup-file.R
\name{post.get.HSPF.sup.file}
\alias{post.get.HSPF.sup.file}
\title{get reads in HSPF-sup-file and returns a long-format data.frame with accum and sqolim info}
\usage{
post.get.HSPF.sup.file(chr.file.sup)
}
\arguments{
\item{chr.file.sup}{is the file name for the HSPF-sup-file (includes path if necessary)}
}
\description{
get reads in HSPF-sup-file and returns a long-format data.frame with accum and sqolim info
}
