% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-get-mutsin-file-names.R
\name{post.get.mutsin.file.names}
\alias{post.get.mutsin.file.names}
\title{gets the names of the mutsin files from the HSPF uci file
returns data.frame with the file names (without path if included) and sub watershed number}
\usage{
post.get.mutsin.file.names(chr.dir.hspf, chr.file.uci)
}
\arguments{
\item{chr.dir.hspf}{path to loaction of hspf input files and where hspf is run}

\item{chr.file.uci}{name of hspf-uci input file}
}
\description{
gets the names of the mutsin files from the HSPF uci file
returns data.frame with the file names (without path if included) and sub watershed number
}
