% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-get-pls-line-info.R
\name{post.get.pls.line.info}
\alias{post.get.pls.line.info}
\title{get pls line info from HSPF-uci-file that correposnds to the HSPF-sup-file info}
\usage{
post.get.pls.line.info(chr.file.uci)
}
\arguments{
\item{chr.file.uci}{is the file name for the HSPF-uci-file (includes path if necessary)}
}
\description{
get pls line info from HSPF-uci-file that correposnds to the HSPF-sup-file info
}
