% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-get-sim-dates.R
\name{post.get.sim.dates}
\alias{post.get.sim.dates}
\title{gets the simulation period from the HSPF uci file
returns data frame with start and end POSIXct dates of simulation period}
\usage{
post.get.sim.dates(chr.dir.hspf, chr.file.uci)
}
\arguments{
\item{chr.dir.hspf}{path to loaction of hspf input files and where hspf is run}

\item{chr.file.uci}{name of hspf-uci input file}
}
\description{
gets the simulation period from the HSPF uci file
returns data frame with start and end POSIXct dates of simulation period
}
