% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-merge-pls-info-to-sup-table.R
\name{post.merge.pls.info.to.sup.table}
\alias{post.merge.pls.info.to.sup.table}
\title{merges HSPF-sup-file and pls line info then returns a long-format data.frame with accum and sqolim info}
\usage{
post.merge.pls.info.to.sup.table(df.pls.line.info, df.sup.table,
  chr.sup.col.num.name)
}
\arguments{
\item{df.pls.line.info}{pls line info that correposnds to the HSPF-sup-file info from HSPF-uci-file}

\item{df.sup.table}{the accum or sqolim loads from the sub-models for the update of sup file}

\item{chr.sup.col.num.name}{the column name in df.sup.table that has the sup-file line numbers for the loads}
}
\description{
merges HSPF-sup-file and pls line info then returns a long-format data.frame with accum and sqolim info
}
