% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-run-model-source.R
\name{post.run.model.source}
\alias{post.run.model.source}
\title{run bacteria source model for a specific input file
Note: this function does not check for the existance of the input file nor
if the input file is the correct format for the source model called. This is
left to the user to ensure.}
\usage{
post.run.model.source(chr.file.input, chr.model)
}
\arguments{
\item{chr.file.input}{is the input file for the model}

\item{chr.model}{is the name of the bacteria source model to use, which is a part of the bacteriaSourceModels package}
}
\description{
run bacteria source model for a specific input file
Note: this function does not check for the existance of the input file nor
if the input file is the correct format for the source model called. This is
left to the user to ensure.
}
