% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-run-model-source-subs.R
\name{post.run.model.source.subs}
\alias{post.run.model.source.subs}
\title{run bacteria source model for a specific input file}
\usage{
post.run.model.source.subs(chr.model, chr.dir.sub.model)
}
\arguments{
\item{chr.model}{is the name of the bacteria source model to use, which is a part of the bacteriaSourceModels}

\item{chr.dir.sub.model}{is the path to the sub model input files for the sub wathersheds}
}
\description{
This function is the bacteria source-model for all the sub watersheds
Note: this function does not check for the existance of the path to the input files nor
if the input files are the correct format for the source model called. This is
left to the user to ensure.
}
