% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-run-models-update-sup-mut.R
\name{post.run.models.update.sup.mut}
\alias{post.run.models.update.sup.mut}
\title{runs models and updates HSPF-sup file (with accum and sqolim info) and MUSTIN loads
returns a list with the updated sup-file as a character vector with each row containing a line of the files and a data.frame for the MUTSIN loads}
\usage{
post.run.models.update.sup.mut(chr.dir.sub.models, df.models, chr.dir.hspf,
  chr.file.sup, chr.file.uci)
}
\arguments{
\item{chr.dir.sub.models}{path to location of sub-model input files}

\item{df.models}{data frame constaining name of sub-mode function and the corresponding path to the model input files
## Not Run:
Example Input
chr.dir.sub.models <- "M:/Models/Bacteria/HSPF/bacteria-sub-model-testing/SourceControl/sub-models"
chr.dir.hspf <- "M:/Models/Bacteria/HSPF/bacteria-sub-model-testing/SourceControl/hspf"
chr.file.sup <- "bigelkwq.sup"
chr.file.uci <- "bigelkwq.uci"

chr.sub.models <- as.character(lsf.str("package:bacteriaSourceModels"))
chr.sub.models <- grep("post", chr.sub.models, invert = TRUE, value = TRUE)

chr.dirs.sub.models <- list.dirs(path = chr.dir.sub.models, full.names = FALSE)
chr.dirs.sub.models <- grep("General", chr.dirs.sub.models[
 nchar(chr.dirs.sub.models) > 0], invert = TRUE, value = TRUE)

df.models <- data.frame(model = chr.sub.models, dir =
                         chr.dirs.sub.models, stringsAsFactors = FALSE)
## End (No run)}

\item{chr.dir.hspf}{path to loaction of hspf input files and where hspf is run}

\item{chr.file.sup}{name of hspf-sup file}

\item{chr.file.uci}{name of hspf-uci input file}
}
\description{
runs models and updates HSPF-sup file (with accum and sqolim info) and MUSTIN loads
returns a list with the updated sup-file as a character vector with each row containing a line of the files and a data.frame for the MUTSIN loads
}
