% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-write-mutsin-files.R
\name{post.write.mutsin.files}
\alias{post.write.mutsin.files}
\title{write updated mutsin files}
\usage{
post.write.mutsin.files(chr.dir.hspf, chr.file.uci, chr.name.wtsd, df.mut)
}
\arguments{
\item{chr.dir.hspf}{path to loaction of hspf input files and where hspf is run}

\item{chr.file.uci}{name of hspf-uci input file}

\item{chr.name.wtsd}{name of the watershed}

\item{df.mut}{data.frame containing mutsin data in the list-class output from function "post.run.models.update.sup.mut".}
}
\description{
write updated mutsin files
}
