% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-write-sup-file.R
\name{post.write.sup.file}
\alias{post.write.sup.file}
\title{write updated hspf-sup file}
\usage{
post.write.sup.file(chr.dir.hspf, chr.file.sup, chr.sup)
}
\arguments{
\item{chr.dir.hspf}{path to loaction of hspf input files and where hspf is run}

\item{chr.file.sup}{name of hspf-sup file to update (WARNING: file is overwrtten)}

\item{chr.sup}{character vector containing lines of the sup file in the list-class output from function "post.run.models.update.sup.mut".}
}
\description{
write updated hspf-sup file
}
