% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wildlife_HerEgr_Sub_Model.R
\name{wildlifeHerEgr}
\alias{wildlifeHerEgr}
\title{model bacteria generation from Hrrons or Egrets}
\usage{
wildlifeHerEgr(chr.input.file)
}
\arguments{
\item{chr.input.file}{is the input file for the model}
}
\description{
This function is the bacteria source-model herons and Ergtrets. The model
generates input for HSPF. The specific outputs from this source model
are loads from the beaver to land and directly to the stream
The load to the land is in the form of load/acre for forest
PLS in a sub-watershed that the source-model contributes to and
the hourly load to the stream in the form of a MUTSIN file. The
input for the model is from an ASCII text file. Use the text below
as a template for the input file. The symbol used for comments in
the input file is "***". The definitions for the symbols used in
the template are: YYYY is four-digit year, MM two-digit month,
DD is the two-digit day, ## is an integer, #.# is a floating point
number, and #.#E+## is a number in scientific notation
}
